port check 055: version_format_is_correct_in_NAME_table
(issue #1388)
