+++ bin/fontbakery-check-ttf.py
@@ -204,7 +204,6 @@ def fontbakery_check_ttf(config):
       checks.check_glyphset_google_vietnamese(fb, target)
       checks.check_glyphset_google_extras(fb, target)
 
-    checks.check_version_format_is_correct_in_NAME_table(fb, font)
     #checks.check_font_has_latest_ttfautohint_applied(fb,
     #                                                 font,
     #                                                 ttfautohint_missing)
